+++ clients/fides-js/src/fides.ts
@@ -182,6 +182,7 @@ _Fides = {
     fidesDisableBanner: false,
     fidesString: null,
     apiOptions: null,
+    fidesTcfGdprApplies: false,
   },
   fides_meta: {},
   identity: {},