+++ clients/fides-js/src/lib/consent-types.ts
@@ -81,6 +81,9 @@ export type FidesOptions = {
 
   // Allows for explicit overrides on various internal API calls made from Fides.
   apiOptions: FidesApiOptions | null;
+
+  // What the "GDPR Applies" field of TCF should default to
+  fidesTcfGdprApplies: boolean;
 };
 
 export type GetPreferencesFnResp = {
@@ -381,11 +384,16 @@ export type OverrideOptions = {
   fides_disable_save_api: boolean;
   fides_disable_banner: boolean;
   fides_embed: boolean;
+  fides_tcf_gdpr_applies: boolean;
 };
 
 export type FidesOptionOverrides = Pick<
   FidesOptions,
-  "fidesString" | "fidesDisableSaveApi" | "fidesEmbed" | "fidesDisableBanner"
+  | "fidesString"
+  | "fidesDisableSaveApi"
+  | "fidesEmbed"
+  | "fidesDisableBanner"
+  | "fidesTcfGdprApplies"
 >;
 
 export type FidesOverrides = {